- Replaced most occurences of "Options" with "Settings" - More work on the Video DVD ripping dialog:   Filename pattern including the beautified volume id or the type of audio stream
svn path=/trunk/extragear/multimedia/k3b/; revision=518009

+++ src/projects/k3bfillstatusdisplay.cpp
@@ -188,7 +188,7 @@ void K3bFillStatusDisplayWidget::paintEvent( QPaintEvent* )
   // draw the text marks
   crect = rect();
   QString text;
-  if( d->cdSize.mode1Bytes() >= (long long)d->doc->size() )
+  if( d->cdSize.mode1Bytes() >= d->doc->size() )
     text = i18n("Available: %1 of %2")
       .arg( d->showTime
 	    ? i18n("%1 min").arg((K3b::Msf( cdSize*60*75 ) - d->doc->length()).toString(false))
